+++ src/components/Footer/Footer.jsx
@@ -1,11 +1,9 @@
 import React from "react";
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 // reactstrap components
 import {
   Button,
-  NavItem,
   NavLink,
-  Nav,
   Container,
   Row,
   Col,
