+++ src/components/Navbars/IndexNavbar.jsx
@@ -105,7 +105,7 @@ class ComponentsNavbar extends React.Component {
               <Row>
                 <Col className="collapse-brand" xs="6">
                   <a href="#pablo" onClick={e => e.preventDefault()}>
-                    BLK•React
+                    See More
                   </a>
                 </Col>
                 <Col className="collapse-close text-right" xs="6">
