+++ src/index.js
@@ -7,25 +7,15 @@ import "assets/scss/blk-design-system-react.scss?v=1.0.0";
 import "assets/demo/demo.css";
 
 import Index from "views/Index.jsx";
-import LandingPage from "views/examples/LandingPage.jsx";
-import RegisterPage from "views/examples/RegisterPage.jsx";
-import ProfilePage from "views/examples/ProfilePage.jsx";
+import ContactPage from "views/examples/ContactPage.jsx";
 
 ReactDOM.render(
   <BrowserRouter>
     <Switch>
       <Route path="/components" render={props => <Index {...props} />} />
       <Route
-        path="/landing-page"
-        render={props => <LandingPage {...props} />}
-      />
-      <Route
-        path="/register-page"
-        render={props => <RegisterPage {...props} />}
-      />
-      <Route
-        path="/profile-page"
-        render={props => <ProfilePage {...props} />}
+        path="/contact-page"
+        render={props => <ContactPage {...props} />}
       />
       <Redirect from="/" to="/components" />
     </Switch>
