+++ src/views/IndexSections/Download.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 // reactstrap components
-import { Button, Container, Row, Col, UncontrolledTooltip } from "reactstrap";
+import { Button, Container, Row, Col } from "reactstrap";
 
 class Download extends React.Component {
   render() {
