+++ src/views/IndexSections/Projects.jsx
@@ -1,0 +1,90 @@
+import React from "react";
+
+// reactstrap components
+import {
+  Button,
+  Container,
+  Row,
+  Col,
+  UncontrolledCarousel
+} from "reactstrap";
+
+const carouselItems = [
+  {
+    src: require("assets/img/denys.jpg"),
+    altText: "Slide 1",
+    caption: ""
+  },
+  {
+    src: require("assets/img/fabien-bazanegue.jpg"),
+    altText: "Slide 2",
+    caption: ""
+  },
+  {
+    src: require("assets/img/mark-finn.jpg"),
+    altText: "Slide 3",
+    caption: ""
+  }
+];
+
+class Projects extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      demoModal: false,
+      miniModal: false,
+      formModal: false
+    };
+  }
+  toggleModal = modalState => {
+    this.setState({
+      [modalState]: !this.state[modalState]
+    });
+  };
+  render() {
+    return (
+      <div className="section section-javascript" id="javascriptComponents">
+        <img alt="..." className="path" src={require("assets/img/path5.png")} />
+        <img
+          alt="..."
+          className="path path1"
+          src={require("assets/img/path5.png")}
+        />
+        <div className="section">
+          <Container>
+            <div className="title">
+              <h2>Sample Works</h2>
+            </div>
+            <Row className="justify-content-between align-items-center">
+              <Col className="mb-5 mb-lg-0" lg="5">
+                <h1 className="text-white font-weight-light">
+                  Bright Events
+                </h1>
+                <p className="text-white mt-4">
+                  A platform that allows Event organizers the freedom to organize and 
+                  manage events. This is built with Python (backend) and ReactJS (frontend).
+                </p>
+                <Button
+                  className="mt-4"
+                  color="warning"
+                  href="https://demos.creative-tim.com/blk-design-system-react/#/documentation/alerts"
+                >
+                  See the application
+                </Button>
+              </Col>
+              <Col lg="6">
+                <UncontrolledCarousel
+                  items={carouselItems}
+                  indicators={false}
+                  autoPlay={false}
+                />
+              </Col>
+            </Row>
+          </Container>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Projects;
