+++ src/views/IndexSections/Tools.jsx
@@ -1,15 +1,7 @@
 import React from "react";
-import classnames from "classnames";
 // reactstrap components
 import {
-  Badge,
-  NavItem,
-  NavLink,
-  Nav,
   Button,
-  Pagination,
-  PaginationItem,
-  PaginationLink,
   Progress,
   Container,
   Row,
